[Refactor] Split large files into focused modules for better code organization
Split router.go (650 lines) into router.go (436 lines) and forwarding.go
(248 lines), separating backend forwarding logic from core routing.

Split metrics.go (568 lines) into metrics.go (152 lines),
endpoint_metrics.go (201 lines), and cluster_metrics.go (262 lines),
grouping metrics by domain: core HTTP, endpoint/backend, and
cluster-level (VIP/BGP/OSPF/TLS/QUIC/policy).

Added filters/doc.go placeholder since the directory was missing.

No functionality changes - identical behavior, improved file structure.

Resolves #25
