+++ internal/agent/filters/doc.go
@@ -1,0 +1,3 @@
+// Package filters is reserved for future filter implementations.
+// Current filter implementations are in the internal/agent/policy package.
+package filters
